Reorganized /DynamicProgramming/ directory and added initial folders/files for Weighted Interval Scheduling
