+++ Guide/DynamicProgramming/Practice/Abbreviation/solution.java
